% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{get_rdata}
\alias{get_rdata}
\title{Read in Rda or Rdata files}
\usage{
get_rdata(data.dir, file.ext = ".Rda")
}
\arguments{
\item{data.dir}{A character string with the name of the directory containing
the files}

\item{file.ext}{An optional character string, defaults to .Rda}
}
\value{
An \code{R} object
}
\description{
\code{get_rdata} reads in all Rda or Rdata files from a directory
}
\details{
This function reads in all Rda or Rdata files in a given directory and saves
them as objects in the Global Environment.
}
\seealso{
\code{\link{load}}
}
