% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{get_rds}
\alias{get_rds}
\title{Read in RDS files}
\usage{
get_rds(data.dir, file.ext = ".Rds")
}
\arguments{
\item{data.dir}{A character string with the name of the directory containing
the RDS files}

\item{file.ext}{An optional character string, defaults to .Rds}
}
\value{
An \code{R} object
}
\description{
\code{get_rds} reads in all RDS files from a directory
}
\details{
This function reads in all RDS files in a given directory and saves them as
objects in the Global Environment.
}
\seealso{
\code{\link{readRDS}}
}

