% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{get_rds_s3}
\alias{get_rds_s3}
\title{Read in RDS files from AWS S3}
\usage{
get_rds_s3(bucket, s3_dir, file_ext = ".Rds")
}
\arguments{
\item{bucket}{character string, bucket to read data from}

\item{s3_dir}{character string, retrieves all files from the specified bucket directory}

\item{file_ext}{An optional character string, defaults to .Rds}
}
\value{
An \code{R} object
}
\description{
\code{get_rds} reads in all RDS files from a directory in an AWS S3 bucket
}
\details{
This function reads in all RDS files in a given directory in an AWS S3 bucket
and saves them as objects in the Global Environment.
}
\seealso{
\code{\link[aws.s3]{s3readRDS}}
}
