% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_files.R
\name{gzip_files}
\alias{gzip_files}
\title{Compress data files}
\usage{
gzip_files(dir.name = "data/raw", pattern = NULL, recursive = FALSE)
}
\arguments{
\item{dir.name}{A string with the name of the directory containing the data
files. Defaults to the \code{data/raw} directory.}

\item{pattern}{An optional regular expression. Only file names which match
the regular expression will be compressed.}

\item{recursive}{A boolean indicating whether to include subdirectories}
}
\description{
\code{gzip_files} compresses all data files in a directory using gzip, if not
already compressed
}
\details{
This function takes the name of a directory and checks all of the data files
within the directory to if they are already compressed. For each file that is
not, it will compress the file using \code{\link[R.utils]{gzip}}.
}
\seealso{
\code{\link[R.utils]{gzip}},
\code{\link[base]{regex}} for regular expressions
}
