% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{make_dirs}
\alias{make_dirs}
\title{Create directories for a data project}
\usage{
make_dirs()
}
\description{
\code{make_dirs} creates a set of standard directories to use for a data
project
}
\details{
This function creates a set of standard directories to organize and
facilitate a data project. The directories created are:

\itemize{
    \item data
    \itemize{
        \item external - any external data, such as data which was manually
        collected
        \item final - final version of the data used for analysis
        \item raw - original, raw data sets
        \item tidy - tidy version of original data sets
    }
    \item doc - manuscript and related documents
    \item explore - exploratory notebooks and figures
    \item figs - final figures for posters or manuscript
    \item irb - IRB protocol and approval documents
    \item report - notebook containing data analysis
    \item src - R script files used to tidy, transform, and aggregate data
    into the final data set
}
}

