% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_rds}
\alias{save_rds}
\title{Save data in RDS files}
\usage{
save_rds(data.dir, pattern, file.ext = ".Rds")
}
\arguments{
\item{data.dir}{A character string with the name of the directory containing
the RDS files}

\item{pattern}{A regular expression indicating the objects to be saved}

\item{file.ext}{An optional character string, defaults to .Rds}
}
\value{
An \code{R} object
}
\description{
\code{save_rds} saves data in RDS files
}
\details{
This function saves all designated data frames in RDS files within a given
directory.
}
\seealso{
\code{\link{saveRDS}}
}

