% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_files.R
\name{ungzip_files}
\alias{ungzip_files}
\title{Uncompress data files}
\usage{
ungzip_files(dir.name = getwd(), pattern = NULL)
}
\arguments{
\item{dir.name}{A string with the name of the directory containing the data
files. Defaults to the current working directory.}

\item{pattern}{An optional regular expression. Only file names which match
the regular expression will be compressed.}
}
\description{
\code{ungzip_files} uncompresses all data files in a directory using gzip
}
\details{
This function takes the name of a directory and uncompresses each file that
is gzipped.
}
\seealso{
\code{\link[R.utils]{gzip}}, \code{\link[base]{regex}} for regular
  expressions
}

