% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_rdata}
\alias{save_rdata}
\title{Save data in Rda or Rdata files}
\usage{
save_rdata(data.dir, pattern, file.ext = ".Rda")
}
\arguments{
\item{data.dir}{A character string with the name of the directory containing
the files}

\item{pattern}{A regular expression indicating the objects to be saved}

\item{file.ext}{An optional character string, defaults to .Rda}
}
\value{
An \code{R} object
}
\description{
\code{save_rdata} saves data in Rda or Rdata files
}
\details{
This function saves all designated data frames in Rda (default) or Rdata
files within a given directory.
}
\seealso{
\code{\link{save}}
}
