% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{save_rds_s3}
\alias{save_rds_s3}
\title{Save file as Rds on AWS S3}
\usage{
save_rds_s3(x, bucket, nm, s3_dir = "data/", encrypt = TRUE)
}
\arguments{
\item{x}{object to be saved, usually a data.frame}

\item{bucket}{character string, bucket to save data to}

\item{nm}{character string with name of object}

\item{s3_dir}{character string, combined with nm to create an object key}

\item{encrypt}{logical, default is to encrypt data}
}
\description{
This function saves a data object as an Rds file on AWS S3. The access key id
and secret access key must be set as environment variables to work.
}
